requirements.txt
google-api-python-client==2.166.0
google-auth==2.40.1
pillow==11.2.1
python-dotenv==1.1.0
streamlit==1.44.1
